Change Log: Fixed issue where saving a campaign before setting sources would turn all sources off.
Added health tracking to the Init tool.

Added Ravenloft Horrors Within as a Source Option.
