+++ src/Runner.java
@@ -18,7 +18,7 @@ import utils.ErrorLogger;
  * Feat Builder
  * 
  */
-
+//TODO: Re-namable players
 public class Runner
 {
 	private static DataContainer data;
